+++ R/future.R
@@ -1,3 +1,103 @@
+#' Make a VM cluster suitable for running parallel workloads
+#' 
+#' This wraps the commands for creating a cluster suitable for \link[future]{future} workloads.
+#' 
+#' @param vm_prefix The prefix of the VMs you want to make. Will be appended the cluster number
+#' @param cluster_size The number of VMs in your cluster
+#' @param docker_image The docker image the jobs on the cluster will run on. Recommend this is derived from \code{rocker/r-parallel}
+#' @param ... Passed to \link{gce_vm_template}
+#' @param ssh_args A list of optional arguments that will be passed to \link{gce_ssh_setup}
+#' 
+#' @export
+#' @import assertthat
+#' @examples 
+#' 
+#' \dontrun{
+#' library(future)
+#' library(googleComputeEngineR)
+#' 
+#' vms <- gce_vm_cluster()
+#' 
+#' ## make a future cluster
+#' plan(cluster, workers = as.cluster(vms))
+#' 
+#' }
+gce_vm_cluster <- function(vm_prefix = "r-cluster-",
+                           cluster_size = 3,
+                           docker_image = "rocker/r-parallel",
+                           ...,
+                           ssh_args = NULL,
+                           project = gce_get_global_project(), 
+                           zone = gce_get_global_zone()){
+  
+  assert_that(
+    is.string(vm_prefix),
+    is.scalar(cluster_size)
+  )
+  
+  ## names for your cluster
+  vm_names <- paste0(vm_prefix, 1:cluster_size)
+  
+  vm_check <- lapply(vm_names, check_vm_exists, project = project, zone = zone)
+  
+  if(all(unlist(lapply(vm_check, is.gce_instance)))){
+    myMessage("VMs already exisits with the cluster names", level = 3)
+    return(vm_check)
+  }
+  
+  dots <- list(...)
+  
+  # defaults
+  dots$name <- NULL
+  dots$template <- "r-base"
+  dots$dynamic_image <- docker_image
+  dots$wait <- FALSE
+  
+  if(is.null(dots$predefined_type)){
+    dots$predefined_type <- "n1-standard-1"
+  }
+  
+  myMessage("# Creating cluster with settings: ", 
+            paste(names(dots), "=", dots, collapse = ", "),
+            level = 3)
+  
+  ## create the cluster using default template for r-base
+  ## creates jobs that are creating VMs in background
+  jobs <- lapply(vm_names, function(x) {
+    dots$name <- x
+    do.call(gce_vm_template, args = dots)
+  })
+  
+  ## wait for all the jobs to complete and VMs are ready
+  jobs <- lapply(jobs, gce_wait)
+  
+  ## get the VM objects
+  vms <- lapply(vm_names, gce_vm)
+  
+  if(!is.null(ssh_args)){
+    assert_that(is.list(ssh_args))
+    myMessage("# Setting up SSH:", paste(names(ssh_args), "=", ssh_args, collapse = ","),
+              level = 3)
+  } else {
+    ssh_args <- list()
+  }
+  
+  vms <- lapply(vms, function(x){
+    ssh_args$instance <- x
+    do.call(gce_ssh_setup, args = ssh_args)
+  }) 
+  
+  myMessage("# Testing cluster:", level = 3)
+  lapply(vms, function(x){
+    gce_ssh(x, paste("echo", x$name, "ssh working"))
+  })
+  
+  vms
+  
+}
+
+
+
 #' Make the Docker cluster on Google Compute Engine
 #' 
 #' Called by \link{as.cluster}
@@ -6,7 +106,7 @@
 #' @param docker_image The docker image to use on the cluster
 #' @param rscript The Rscript command to run on the cluster
 #' @param rscript_args Arguments to the RScript
-#' @param install_future Whether to check if future is installed first 
+#' @param install_future Whether to check if future is installed first (not needed if using docker derived from rocker/r-parallel which is recommended)
 #' @param ... Other arguments passed to \link[future]{makeClusterPSOCK}
 #' @param verbose How much feedback to show
 #' 
@@ -14,15 +114,15 @@
 #' @author Henrik Bengtsson \email{henrikb@@braju.com}
 #' @export
 makeDockerClusterPSOCK <- function(workers, 
-                                   docker_image = "rocker/r-base", 
+                                   docker_image = "rocker/r-parallel", 
                                    rscript = c("docker", "run", "--net=host", docker_image, "Rscript"), 
-                                   rscript_args = NULL, install_future = TRUE, ..., verbose = FALSE) {
+                                   rscript_args = NULL, install_future = FALSE, ..., verbose = FALSE) {
   ## Should 'future' package be installed, if not already done?
   if (install_future) {
     rscript_args <- c("-e", 
-                      shQuote(sprintf("if (!requireNamespace('future', quietly = TRUE)) install.packages('future', quiet = %s)", 
+      shQuote(sprintf("if (!requireNamespace('future', quietly = TRUE)) install.packages('future', quiet = %s)", 
                                       !verbose)), 
-                      rscript_args)
+      rscript_args)
   }
   
   makeClusterPSOCK(workers, 
@@ -77,70 +177,5 @@ as.cluster.gce_instance <- function(x,
   stopifnot(!is.null(ips))
   
   makeDockerClusterPSOCK(ips, user = x$ssh$username, rshopts = rshopts, ...)
-}
-
-
-#' Install R packages onto an instance's stopped docker image
-#' 
-#' @param instance The instance running the container
-#' @param docker_image A docker image to install packages within.
-#' @param cran_packages A character vector of CRAN packages to be installed
-#' @param github_packages A character vector of devtools packages to be installed
-#' 
-#' @details 
-#' 
-#' See the images on the instance via \code{docker_cmd(instance, "images")}
-#' 
-#' If using devtools github, will look for an auth token via \code{devtools::github_pat()}.  
-#'   This is an environment variable called \code{GITHUB_PAT} 
-#' 
-#'  Will start a container, install packages and then commit 
-#'    the container to an image of the same name via \code{docker commit -m "installed packages via gceR"}
-#' 
-#' @return TRUE if successful
-#' @import future
-#' @importFrom utils install.packages
-#' @importFrom remotes install_github
-#' @export
-gce_future_install_packages <- function(instance,
-                                        docker_image,
-                                        cran_packages = NULL,
-                                        github_packages = NULL){
-  
-  ## set up future cluster
-  temp_name <- paste0("gceR-install-",idempotency())
-  clus <- as.cluster(instance, 
-                             docker_image = docker_image,
-                             rscript = c("docker", "run",paste0("--name=",temp_name),"--net=host", docker_image, "Rscript"))
-  
-  plan(cluster, workers = clus)
-  
-  if(!is.null(cran_packages)){
-    cran <- NULL
-    cran %<-% install.packages(cran_packages)
-    cran
-  }
-  
-  if(!is.null(github_packages)){
-    devt <- NULL
-    devt %<-% install_github(github_packages, auth_token = github_pat())
-    devt
-  }
-  
-  docker_cmd(instance, 
-                     cmd = "commit", 
-                     args = c("-a 'googleComputeEngineR'" ,
-                              paste("-m 'Installed packages:", 
-                                    paste(collapse = " ", cran_packages), 
-                                    paste(collapse = " ", github_packages),
-                                    "'"),
-                              temp_name, 
-                              docker_image))
-  
-  ## stop the container
-  docker_cmd(instance, "stop", temp_name)
-  
-  TRUE
-  
 }